requirements.txt
mysqlclient==2.1.1
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2